Re #27226 Move sub-presenter init code out of UI
Various sub-presenters were being initialized in the UI file using
various factories. This simplifies the code by removing said factories,
and moving the code into add_runs_presenter, who previously called it
anyway.
